+++ src/libopensc/internal.h
@@ -157,15 +157,15 @@ int _sc_match_atr(struct sc_card *card, const struct sc_atr_table *table, int *t
 int _sc_card_add_algorithm(struct sc_card *card, const struct sc_algorithm_info *info);
 int _sc_card_add_symmetric_alg(sc_card_t *card, unsigned int algorithm,
 			       unsigned int key_length, unsigned long flags);
-int _sc_card_add_rsa_alg(struct sc_card *card, unsigned int key_length,
+int _sc_card_add_rsa_alg(struct sc_card *card, size_t key_length,
 		unsigned long flags, unsigned long exponent);
-int _sc_card_add_ec_alg(struct sc_card *card, unsigned int key_length,
+int _sc_card_add_ec_alg(struct sc_card *card, size_t key_length,
 		unsigned long flags, unsigned long ext_flags,
 		struct sc_object_id *curve_oid);
-int _sc_card_add_eddsa_alg(struct sc_card *card, unsigned int key_length,
+int _sc_card_add_eddsa_alg(struct sc_card *card, size_t key_length,
 		unsigned long flags, unsigned long ext_flags,
 		struct sc_object_id *curve_oid);
-int _sc_card_add_xeddsa_alg(struct sc_card *card, unsigned int key_length,
+int _sc_card_add_xeddsa_alg(struct sc_card *card, size_t key_length,
 		unsigned long flags, unsigned long ext_flags,
 		struct sc_object_id *curve_oid);
 
@@ -224,7 +224,7 @@ int sc_get_encoding_flags(sc_context_t *ctx,
 int sc_mutex_create(const sc_context_t *ctx, void **mutex);
 /**
  * Tries to acquire a lock for a sc_mutex object. Note: Unless
- * sc_mutex_set_mutex_funcs() has been called before this 
+ * sc_mutex_set_mutex_funcs() has been called before this
  * function does nothing and always returns SUCCESS.
  * @param  ctx    sc_context_t object with the thread context
  * @param  mutex  mutex object to lock
